// webapp/src/types/correcteur.ts
export interface Correcteur {
    idCorrecteur: string;
    nom: string;
    prenom: string;
    immatricule: string;
    cin: string;
    adresse: string;
    adresseProfession: string;
    telephone: string;
    specialite: string;
    secteur: string;
    option: string;
    matiere: string;
    grade: string;
    statut: string;
}